hw9 done without slices
